feat: migrate formUtils.spec
